+++ VSProject/VSProject/Global.cpp
@@ -3,19 +3,19 @@
 namespace Global
 {
 	//time
-	double deltaTime = 0;
+	float deltaTime = 0;
 	const double fixedDeltaTime = 1000.0 / 100.0; //basically number of physics updates per second
 
 	//physics
-	const Vector3 gravity(0, -9.81, 0);
-	const double gravityMag = -9.81;
-	const double airDensity = 1.225;
+	const Vector3 gravity(0, -9.81f, 0);
+	const float gravityMag = -9.81f;
+	const float airDensity = 1.225f;
 
 	bool shouldUpdate = true;
 
 	bool writeContactDataToFile = false;
 
-	const Vector3 skyColour = Vector3(0.529, 0.808, 0.922);
+	const Vector3 skyColour = Vector3(0.529f, 0.808f, 0.922f);
 }
 
 //double Global::GetDeltaTime()
